Apply Rider's Full Cleanup layout
Reformat with Rider's Full Cleanup, the layout authority for the KillerPixelCrew repositories, and turn off Roslyn's IDE0055 where the .editorconfig enabled it. The unsafe modifiers the cleanup removed are restored; the inspection is wrong for the SDK's language version.

+++ tests/WindowsDeviceControl.Tests/PowerRequestTests.cs
@@ -13,13 +13,18 @@ public sealed class PowerRequestTests
         FakeApi api = new();
         var request = new WindowsPowerRequest("test reason", WindowsPowerRequestKind.System, api);
         Assert.Equal(0, api.Creates);
-        request.Acquire(); request.Acquire();
+        request.Acquire();
+        request.Acquire();
         Assert.True(request.IsHeld);
-        Assert.Equal(1, api.Creates); Assert.Equal(1, api.Sets);
+        Assert.Equal(1, api.Creates);
+        Assert.Equal(1, api.Sets);
         Assert.Equal("test reason", Marshal.PtrToStringUni(api.Reason));
-        request.Release(); request.Release();
-        Assert.False(request.IsHeld); Assert.Equal(1, api.Clears);
-        request.Dispose(); request.Dispose();
+        request.Release();
+        request.Release();
+        Assert.False(request.IsHeld);
+        Assert.Equal(1, api.Clears);
+        request.Dispose();
+        request.Dispose();
         Assert.Equal(1, api.Closes);
         Assert.Throws<ObjectDisposedException>(() => request.Acquire());
     }
@@ -34,7 +39,8 @@ public sealed class PowerRequestTests
         Assert.True(request.IsHeld);
         request.Dispose();
         Assert.False(request.IsHeld);
-        Assert.Equal(1, api.Clears); Assert.Equal(1, api.Closes);
+        Assert.Equal(1, api.Clears);
+        Assert.Equal(1, api.Closes);
     }
 
     [Fact]
@@ -43,7 +49,8 @@ public sealed class PowerRequestTests
         FakeApi api = new() { SetSucceeds = false };
         using var request = new WindowsPowerRequest("test", WindowsPowerRequestKind.System, api);
         Assert.Throws<Win32Exception>(() => request.Acquire());
-        Assert.False(request.IsHeld); Assert.Equal(1, api.Sets);
+        Assert.False(request.IsHeld);
+        Assert.Equal(1, api.Sets);
     }
 
     private sealed class FakeApi : IPowerRequestApi
@@ -52,9 +59,30 @@ public sealed class PowerRequestTests
         internal nint Reason;
         internal bool SetSucceeds = true, ClearSucceeds = true;
         public int LastError => 5;
-        public nint Create(nint reason) { Creates++; Reason = reason; return 17; }
-        public bool Set(nint request, int kind) { Sets++; return SetSucceeds; }
-        public bool Clear(nint request, int kind) { Clears++; return ClearSucceeds; }
-        public void Close(nint request) { Assert.Equal((nint)17, request); Closes++; }
+
+        public nint Create(nint reason)
+        {
+            Creates++;
+            Reason = reason;
+            return 17;
+        }
+
+        public bool Set(nint request, int kind)
+        {
+            Sets++;
+            return SetSucceeds;
+        }
+
+        public bool Clear(nint request, int kind)
+        {
+            Clears++;
+            return ClearSucceeds;
+        }
+
+        public void Close(nint request)
+        {
+            Assert.Equal(17, request);
+            Closes++;
+        }
     }
 }
